Remove fp.h (#21)
This was replaced by medleyfp.h. Perhaps this happened because
some systems have an fp.h in their system includes and the naming
collision was an issue. But nothing uses fp.h and medleyfp.h is
an updated version of the contents.

This also updates a bunch of manually managed dependency lists
to reflect the usage of medleyfp.h rather than fp.h.

Nick Briggs: Note that on OSF1 medleyfp.h includes the system fp.h (probably the origin of the renaming)