initkbd: Fix warning about missing decl for strcmp.
Instead of adding more stuff to the `#ifdef` around the inclusion
of `<string.h>`, we remove that `#ifdef` entirely.

There are many instances of `<string.h>` being included in the
code without any `#ifdef` protection, and `strcmp` is defined to
be declared in `<string.h>`, so all platforms from the last 20
years and into the future will need this.