Correct async I/O setup to work for Solaris
Solaris does not define O_ASYNC, but does define FASYNC and FNDELAY,
which it expects to be used in place of O_ASYNC in an fcntl() to set up
asynchronous I/O on a socket.

This fix also generates a compiler warning if neither the O_ASYNC nor
FASYNC/FNDELAY methods are available.