Remove `USHORT` typedef.
This was defined as `unsigned` for DOS and `unsigned short` for
everyone else, but it was only used in bitfield definitions
where just having `unsigned` is fine.

Other bitfield definitions (in some cases, immediately adjacent
to the ones modified here), were already just using a bare
`unsigned`, so this improves the consistency of the definitions.